Move the API to an new file
