+++ jsonschema_gentypes/api_draft_04.py
@@ -1,0 +1,430 @@
+"""
+The API version draft 04.
+"""
+
+import re
+from typing import Dict, List, Union, cast
+
+from jsonschema_gentypes import (
+    BuiltinType,
+    CombinedType,
+    NativeType,
+    Type,
+    TypedDictType,
+    TypeEnum,
+    configuration,
+    get_description,
+    get_name,
+    jsonschema_draft_04,
+    jsonschema_draft_06,
+    jsonschema_draft_07,
+    jsonschema_draft_2019_09,
+)
+from jsonschema_gentypes.api import API
+
+
+class APIv4(API):
+    """
+    JSON Schema draft 4.
+    """
+
+    def const(
+        self,
+        schema: Union[
+            jsonschema_draft_04.JSONSchemaD4,
+            jsonschema_draft_06.JSONSchemaItemD6,
+            jsonschema_draft_07.JSONSchemaItemD7,
+            jsonschema_draft_2019_09.JSONSchemaItemD2019,
+        ],
+        proposed_name: str,
+    ) -> Type:
+        """
+        Generate a ``Literal`` for a const value.
+        """
+        raise NotImplementedError("const is not supported in draft 4")
+
+    def enum(
+        self,
+        schema: Union[
+            jsonschema_draft_04.JSONSchemaD4,
+            jsonschema_draft_06.JSONSchemaItemD6,
+            jsonschema_draft_07.JSONSchemaItemD7,
+            jsonschema_draft_2019_09.JSONSchemaItemD2019,
+        ],
+        proposed_name: str,
+    ) -> Type:
+        """
+        Generate an enum.
+        """
+        return TypeEnum(
+            get_name(schema, proposed_name),
+            cast(List[Union[int, float, bool, str, None]], schema["enum"]),
+            get_description(schema),
+        )
+
+    def boolean(
+        self,
+        schema: Union[
+            jsonschema_draft_04.JSONSchemaD4,
+            jsonschema_draft_06.JSONSchemaItemD6,
+            jsonschema_draft_07.JSONSchemaItemD7,
+            jsonschema_draft_2019_09.JSONSchemaItemD2019,
+        ],
+        proposed_name: str,
+    ) -> Type:
+        """
+        Generate a ``bool`` annotation for a boolean object.
+        """
+        del schema, proposed_name
+        return BuiltinType("bool")
+
+    def object(
+        self,
+        schema: Union[
+            jsonschema_draft_04.JSONSchemaD4,
+            jsonschema_draft_06.JSONSchemaItemD6,
+            jsonschema_draft_07.JSONSchemaItemD7,
+            jsonschema_draft_2019_09.JSONSchemaItemD2019,
+        ],
+        proposed_name: str,
+    ) -> Type:
+        """
+        Generate an annotation for an object, usually a TypedDict.
+        """
+
+        std_dict = None
+        name = get_name(schema, proposed_name)
+        additional_properties = cast(
+            Union[
+                jsonschema_draft_04.JSONSchemaD4,
+                jsonschema_draft_06.JSONSchemaD6,
+                jsonschema_draft_07.JSONSchemaD7,
+                jsonschema_draft_2019_09.JSONSchemaD2019,
+            ],
+            schema.get("additionalProperties"),
+        )
+        if (
+            additional_properties is True
+            and self.additional_properties == configuration.ADDITIONALPROPERTIES_ALWAYS
+        ):
+            std_dict = CombinedType(NativeType("Dict"), [BuiltinType("str"), NativeType("Any")])
+        elif isinstance(additional_properties, dict):
+            sub_type = self.get_type(additional_properties, f"{proposed_name} additionalProperties")
+            std_dict = CombinedType(NativeType("Dict"), [BuiltinType("str"), sub_type])
+        properties = cast(
+            Dict[
+                str,
+                Union[
+                    jsonschema_draft_04.JSONSchemaD4,
+                    jsonschema_draft_06.JSONSchemaItemD6,
+                    jsonschema_draft_07.JSONSchemaItemD7,
+                    jsonschema_draft_2019_09.JSONSchemaItemD2019,
+                ],
+            ],
+            schema.get("properties"),
+        )
+        proposed_name = schema.get("title", proposed_name)
+        if properties:
+            required = set(schema.get("required", []))
+
+            struct = {
+                prop: self.get_type(sub_schema, proposed_name + " " + prop, auto_alias=False)
+                for prop, sub_schema in properties.items()
+            }
+
+            type_: Type = TypedDictType(
+                name if std_dict is None else name + "Typed",
+                struct,
+                get_description(schema) if std_dict is None else [],
+                required=required,
+            )
+
+            comments = []
+
+            if std_dict is not None:
+                type_ = CombinedType(NativeType("Union"), [std_dict, type_])
+                comments += [
+                    "",
+                    "WARNING: Normally the types should be a mix of each other instead of Union.",
+                    "See: https://github.com/camptocamp/jsonschema-gentypes/issues/7",
+                ]
+
+            type_.set_comments(comments)
+            return type_
+        if std_dict is not None:
+            return std_dict
+        return CombinedType(NativeType("Dict"), [BuiltinType("str"), NativeType("Any")])
+
+    def array(
+        self,
+        schema: Union[
+            jsonschema_draft_04.JSONSchemaD4,
+            jsonschema_draft_06.JSONSchemaItemD6,
+            jsonschema_draft_07.JSONSchemaItemD7,
+            jsonschema_draft_2019_09.JSONSchemaItemD2019,
+        ],
+        proposed_name: str,
+    ) -> Type:
+        """
+        Generate a ``List[]`` annotation with the allowed types.
+        """
+        items = schema.get("items")
+        if items is True:
+            return CombinedType(NativeType("List"), [NativeType("Any")])
+        elif items is False:
+            raise NotImplementedError('"items": false is not supported')
+        elif isinstance(items, list):
+            inner_types = [
+                self.get_type(
+                    cast(
+                        Union[
+                            jsonschema_draft_04.JSONSchemaD4,
+                            jsonschema_draft_06.JSONSchemaItemD6,
+                            jsonschema_draft_07.JSONSchemaItemD7,
+                            jsonschema_draft_2019_09.JSONSchemaItemD2019,
+                        ],
+                        item,
+                    ),
+                    f"{proposed_name} {nb}",
+                )
+                for nb, item in enumerate(items)
+            ]
+            type_: Type = CombinedType(NativeType("Tuple"), inner_types)
+            if {schema.get("minItems"), schema.get("maxItems")} - {None, len(items)}:
+                type_.set_comments(
+                    [
+                        "WARNING: 'items': If list, must have minItems == maxItems.",
+                        "See: https://json-schema.org/understanding-json-schema/"
+                        "reference/array.html#tuple-validation",
+                    ]
+                )
+                return type_
+            return type_
+        elif items is not None:
+            return CombinedType(
+                NativeType("List"),
+                [
+                    self.get_type(
+                        cast(
+                            Union[
+                                jsonschema_draft_04.JSONSchemaD4,
+                                jsonschema_draft_06.JSONSchemaItemD6,
+                                jsonschema_draft_07.JSONSchemaItemD7,
+                                jsonschema_draft_2019_09.JSONSchemaItemD2019,
+                            ],
+                            items,
+                        ),
+                        proposed_name + " item",
+                    )
+                ],
+            )
+        else:
+            type_ = BuiltinType("None")
+            type_.set_comments(["WARNING: we get an array without any items"])
+            return type_
+
+    def any_of(
+        self,
+        schema: Union[
+            jsonschema_draft_04.JSONSchemaD4,
+            jsonschema_draft_06.JSONSchemaItemD6,
+            jsonschema_draft_07.JSONSchemaItemD7,
+            jsonschema_draft_2019_09.JSONSchemaItemD2019,
+        ],
+        sub_schema: List[
+            Union[
+                jsonschema_draft_04.JSONSchemaD4,
+                jsonschema_draft_06.JSONSchemaItemD6,
+                jsonschema_draft_07.JSONSchemaItemD7,
+                jsonschema_draft_2019_09.JSONSchemaItemD2019,
+            ]
+        ],
+        proposed_name: str,
+        sub_name: str,
+    ) -> Type:
+        """
+        Generate a ``Union`` annotation with the allowed types.
+        """
+        inner_types = list(
+            filter(
+                lambda o: o is not None,
+                [
+                    self.get_type(subs, f"{proposed_name} {sub_name}{index}")
+                    for index, subs in enumerate(sub_schema)
+                ],
+            )
+        )
+        return CombinedType(NativeType("Union"), inner_types)
+
+    def ref_to_proposed_name(self, ref: str) -> str:
+        """
+        Create a proposed name from a ref.
+
+        Change the case from camel case to standard case, this is for layer inverse conversion then don't lost the case.
+        """
+        ref_proposed_name = ref
+        if ref.startswith("#/definitions/"):
+            ref_proposed_name = ref[len("#/definitions/") :]
+        elif ref.startswith("#/"):
+            ref_proposed_name = ref[len("#/") :]
+        if "/" in ref_proposed_name:
+            ref_proposed_name = ref_proposed_name.replace("/", " ")
+        else:
+            if re.search("[a-z]", ref_proposed_name):
+                ref_proposed_name = re.sub("([a-z0-9])([A-Z])", r"\1 \2", ref_proposed_name).lower()
+        return ref_proposed_name
+
+    def ref(
+        self,
+        schema: Union[
+            jsonschema_draft_04.JSONSchemaD4,
+            jsonschema_draft_06.JSONSchemaItemD6,
+            jsonschema_draft_07.JSONSchemaItemD7,
+            jsonschema_draft_2019_09.JSONSchemaItemD2019,
+        ],
+        proposed_name: str,
+    ) -> Type:
+        """
+        Handle a `$ref`.
+        """
+
+        # ref is not correctly declared in draft 4.
+        schema_casted = cast(
+            Union[
+                jsonschema_draft_06.JSONSchemaItemD6,
+                jsonschema_draft_07.JSONSchemaItemD7,
+                jsonschema_draft_2019_09.JSONSchemaItemD2019,
+            ],
+            schema,
+        )
+
+        ref = schema_casted["$ref"]
+        del schema_casted["$ref"]
+
+        if ref == "#":  # Self ref.
+            # Per @ilevkivskyi:
+            #
+            # > You should never use ForwardRef manually
+            # > Also it is deprecated and will be removed soon
+            # > Support for recursive types is limited to proper classes
+            # > currently
+            #
+            # forward_ref = ForwardRef(UnboundType(self.outer_name))
+            # self.forward_refs.append(forward_ref)
+            # return forward_ref
+
+            assert self.root is not None
+            return self.root
+
+        if ref in self.ref_type:
+            return self.ref_type[ref]
+
+        resolve = getattr(self.resolver, "resolve", None)
+        if resolve is None:
+            with self.resolver.resolving(ref) as resolved:
+                schema_casted.update(resolved)
+                type_ = self.get_type(schema_casted, self.ref_to_proposed_name(ref))
+        else:
+            scope, resolved = self.resolver.resolve(ref)
+            self.resolver.push_scope(scope)
+            try:
+                schema_casted.update(resolved)
+                type_ = self.get_type(schema_casted, self.ref_to_proposed_name(ref))
+            finally:
+                self.resolver.pop_scope()
+
+        if ref:
+            self.ref_type[ref] = type_
+        return type_
+
+    def string(
+        self,
+        schema: Union[
+            jsonschema_draft_04.JSONSchemaD4,
+            jsonschema_draft_06.JSONSchemaItemD6,
+            jsonschema_draft_07.JSONSchemaItemD7,
+            jsonschema_draft_2019_09.JSONSchemaItemD2019,
+        ],
+        proposed_name: str,
+    ) -> Type:
+        """
+        Generate a ``str`` annotation.
+        """
+        del schema, proposed_name
+        return BuiltinType("str")
+
+    def number(
+        self,
+        schema: Union[
+            jsonschema_draft_04.JSONSchemaD4,
+            jsonschema_draft_06.JSONSchemaItemD6,
+            jsonschema_draft_07.JSONSchemaItemD7,
+            jsonschema_draft_2019_09.JSONSchemaItemD2019,
+        ],
+        proposed_name: str,
+    ) -> Type:
+        """
+        Generate a ``Union[int, float]`` annotation.
+        """
+        del schema, proposed_name
+        return CombinedType(NativeType("Union"), [BuiltinType("int"), BuiltinType("float")])
+
+    def integer(
+        self,
+        schema: Union[
+            jsonschema_draft_04.JSONSchemaD4,
+            jsonschema_draft_06.JSONSchemaItemD6,
+            jsonschema_draft_07.JSONSchemaItemD7,
+            jsonschema_draft_2019_09.JSONSchemaItemD2019,
+        ],
+        proposed_name: str,
+    ) -> Type:
+        """
+        Generate an ``int`` annotation.
+        """
+        del schema, proposed_name
+        return BuiltinType("int")
+
+    def null(
+        self,
+        schema: Union[
+            jsonschema_draft_04.JSONSchemaD4,
+            jsonschema_draft_06.JSONSchemaItemD6,
+            jsonschema_draft_07.JSONSchemaItemD7,
+            jsonschema_draft_2019_09.JSONSchemaItemD2019,
+        ],
+        proposed_name: str,
+    ) -> Type:
+        """
+        Generate an ``None`` annotation.
+        """
+        del schema, proposed_name
+        return BuiltinType("None")
+
+    def default(
+        self,
+        schema: Union[
+            jsonschema_draft_04.JSONSchemaD4,
+            jsonschema_draft_06.JSONSchemaItemD6,
+            jsonschema_draft_07.JSONSchemaItemD7,
+            jsonschema_draft_2019_09.JSONSchemaItemD2019,
+        ],
+        proposed_name: str,
+    ) -> Type:
+        """
+        Treat the default keyword.
+
+        See: https://json-schema.org/understanding-json-schema/reference/generic.html
+        """
+
+        type_ = "Any"
+        for test_type, type_name in [
+            (str, "str"),
+            (int, "int"),
+            (float, "float"),
+            (bool, "bool"),
+        ]:
+            if isinstance(schema["default"], test_type):
+                type_ = type_name
+        the_type = BuiltinType(type_)
+        return the_type
