+++ jsonschema_gentypes/api_draft_07.py
@@ -1,0 +1,11 @@
+"""
+The API version draft 07.
+"""
+
+from jsonschema_gentypes.api_draft_06 import APIv6
+
+
+class APIv7(APIv6):
+    """
+    JSON Schema draft 7.
+    """
